fix: forcefully reset grid/generator info if they are absent
I noticed that generator/grid info was staying non-zero/non-unknown
even after the generator/grid were absent. I did not take the time
to find out whether it was because the cube was sending old data,
or if it was because it was not sending data anymore at all.

Instead, I made sure we handle both cases. I think that given the
track record of whisperpower, we're safer that way.